scrapi/consumers/figshare/requirements.txt
requests==2.4.1
nameparser==0.3.3
python-dateutil==2.2
